requirements.txt
# local package
-e .
awscli

# external requirements
click
coverage
flake8
matplotlib
numpy

pre-commit
python-dotenv>=0.5.1
scikit-learn
scipy
Sphinx
torch
